Reinventing the wheel once more for CLI tools structure
Got tired of all the 'duplicated code' PEP warnings in PyCharm, so now all CLI tools will be implemented by subclassing the 'CLITool' superclass in the commons package, which is now called 'commons' instead of 'common-utils'. Also, default values are now read from a configuration file ($HOME/.aws-cli-utils) so that everyone can use this and not just the creator, which was stupid. This commit changes:
 - The EC2 command is now implemented using the inheritance structure
 - The EC2 command now takes default values from the configuration file. The key is 'EC2' and users can provide values for 'DefaultInstanceName', 'DefaultUsername' and 'DefaultPort' under it
 - The 'common-utils' package is now called 'commons' for better clarity